change type of p into int16_t

+++ in4073/drone.h
@@ -27,10 +27,10 @@ char cur_mode;
 char prev_mode;
 
 //p controller value
-int8_t p;
-int8_t p1;
-int8_t p2;
-int8_t p3;
+int16_t p;
+int16_t p1;
+int16_t p2;
+int16_t p3;
 
 //variable to hold current movement
 int8_t cur_lift;
